refactor: move fs methods to utils
